+++ todo
@@ -2,6 +2,7 @@ Current:
 Check how edit works with new multistep form
 
 To do
+Preview certificate
 Reuse images uploaded for a session on another slide - have a add a previously uploaded image button which launches gallery and adds the src to the uploader preview array
 Indicate the default questions on feedback create
 Static slide content overlapping ediges on mobile (see screenshot on AM teams convo)
